requirements.txt
absl-py==2.1.0
altair==5.3.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
astropy==6.0.1
astropy-iers-data==0.2024.4.22.0.29.50
astunparse==1.6.3
attrs==23.2.0
bidict==0.23.1
biosppy==2.2.1
blinker==1.8.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.2
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.61.1
django-cors-headers==4.3.1
django-filter==24.2
django-oauth-toolkit==2.3.0
djangorestframework==3.15.1
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.3.1
Faker==22.4.0
filelock==3.13.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
future==1.0.0
gast==0.5.4
gatspy==0.3
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
heartpy==1.2.7
HeatPy==0.1
hrv-analysis==1.0.4
html5lib==1.1
huggingface-hub==0.22.2
idna==3.6
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mock==5.1.0
namex==0.0.8
nltk==3.8.1
nolds==0.5.2
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
oscrypto==1.3.0
packaging==24.0
pandas==2.2.2
pathlib2==2.3.7.post1
pdfkit==1.0.0
PeakUtils==1.3.4
pillow==10.2.0
protobuf==4.25.3
psycopg2==2.9.9
py-ecg-detectors==1.3.5
pyarrow==16.0.0
pycparser==2.21
pydeck==0.9.0b1
pyerfa==2.0.1.4
Pygments==2.17.2
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.0.1
pypng==0.20220715.0
pysiology==0.0.9.6
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.35.0
regex==2024.4.16
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setuptools==69.5.1
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
streamlit==1.33.0
svglib==1.5.1
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.4.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
transformers==4.40.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.0
watchdog==4.0.0
webencodings==0.5.1
Werkzeug==3.0.2
wheel==0.43.0
wkhtmltopdf==0.2
wrapt==1.16.0
xgboost==2.0.3
xhtml2pdf==0.2.14
